fix memory leaks caused by smart-ptr cross links
The fix of a large memory leak (commit 825e383be7f0407d397be24573761ea2fe055b61) and
the transition to nodes life management via smart-pointers led to the problem of leakage
through cross-references. It is unclear how to solve this problem without rewriting
the library code.
For now, a crutch-solution is being used in the form of the destroy_cross_references()
function.